Add taxonomy term view. Close #74.
The taxonomy term view is based on the appearance of the search view.
There is a known bug in Drupal 7 taxonomy-term.tpl.php that causes this
template to work in a completely different way than the rest of themplates.
This template only acts as a 'header' and the node teasers are automatically
rendered used the node template.  The solution was to edit the node templates
to check if the view mode is 'full' or 'teaser' and render the full page
or the resumed one in each case.